@@ -36,91 +36,46 @@
     "yarn": ">= 1.0.0"
   },
   "dependencies": {
-    "@angular/animations": "^8.1.1",
-    "@angular/cdk": "^8.1.1",
-    "@angular/common": "^8.1.1",
-    "@angular/compiler": "^8.1.1",
-    "@angular/core": "^8.1.1",
-    "@angular/flex-layout": "^8.0.0-beta.26",
-    "@angular/forms": "^8.1.1",
-    "@angular/material": "^8.1.1",
-    "@angular/material-moment-adapter": "^8.1.1",
-    "@angular/platform-browser": "^8.1.1",
-    "@angular/platform-browser-dynamic": "^8.1.1",
-    "@angular/router": "^8.1.1",
-    "@fullcalendar/core": "^4.1.0",
-    "@fullcalendar/daygrid": "^4.1.0",
-    "@fullcalendar/list": "^4.1.0",
-    "@fullcalendar/moment": "^4.2.0",
-    "@fullcalendar/moment-timezone": "^4.2.0",
-    "@fullcalendar/timegrid": "^4.1.0",
+    "@angular/animations": "^8.2.7",
+    "@angular/cdk": "^8.2.0",
+    "@angular/common": "^8.2.7",
+    "@angular/compiler": "^8.2.7",
+    "@angular/core": "^8.2.7",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.7",
+    "@angular/material": "^8.2.0",
+    "@angular/material-moment-adapter": "^8.2.0",
+    "@angular/platform-browser": "^8.2.7",
+    "@angular/platform-browser-dynamic": "^8.2.7",
+    "@angular/router": "^8.2.7",
     "@mat-datetimepicker/core": "^3.0.0-beta.0",
     "@mat-datetimepicker/moment": "^2.0.1",
-    "@material-ui/codemod": "^4.0.2",
-    "@material-ui/core": "^4.0.2",
-    "@material-ui/icons": "^4.0.1",
-    "@material-ui/lab": "^4.0.0-alpha.15",
-    "@material-ui/styles": "^4.0.2",
-    "@material-ui/system": "^4.0.2",
-    "@material-ui/types": "^4.0.1",
-    "@material-ui/utils": "^4.0.1",
-    "@types/lodash": "^4.14.134",
-    "angular": "^1.7.8",
-    "angular-animate": "^1.7.8",
-    "angular-aria": "^1.7.8",
-    "angular-chart.js": "^1.1.1",
-    "angular-drag-and-drop-lists": "^2.1.0",
-    "angular-froala-wysiwyg": "2.7.1",
-    "angular-material": "^1.1.19",
-    "angular-messages": "^1.7.8",
-    "angular-paging": "^2.2.2",
-    "angular-resource": "^1.7.8",
-    "angular-sanitize": "^1.7.8",
-    "angular-scroll-spy": "^1.0.0",
-    "angular-ui-tree": "^2.22.6",
+    "@material-ui/codemod": "^4.3.0",
+    "@material-ui/core": "^4.4.2",
+    "@material-ui/icons": "^4.4.1",
+    "@material-ui/lab": "^4.0.0-alpha.26",
+    "@material-ui/styles": "^4.4.1",
+    "@material-ui/system": "^4.3.3",
+    "@material-ui/types": "^4.1.1",
+    "@material-ui/utils": "^4.4.0",
+    "@types/lodash": "^4.14.138",
     "bowser": "^1.9.4, <2.0",
-    "chart.js": "^2.8.0",
-    "codemirror": "^5.47.0",
-    "coffeescript": "^2.4.1",
-    "core-js": "^3.1.3",
-    "froala-editor": "2.7.1",
+    "core-js": "^3.2.1",
     "hammerjs": "^2.0.8",
-    "highlight.js": "^9.15.8",
-    "ical.js": "^1.3.0",
     "jquery": "^3.4.1",
-    "js-md5": "^0.7.3",
-    "less": "^3.9.0",
-    "lodash": "^4.17.11",
-    "masonry-layout": "^4.2.2",
-    "moment": "^2.24.0",
-    "moment-range": "^4.0.2",
-    "moment-timezone": "^0.5.25",
-    "ng-file-upload": "^12.2.13",
-    "ng-material-datetimepicker": "^1.18.5",
-    "ngx-quill": "^7.1.0",
-    "promise-polyfill": "^8.1.0",
-    "quill": "^1.3.6",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "^3.0.1",
-    "remove": "^0.1.5",
-    "requirejs": "^2.3.6",
-    "requirejs-text": "^2.0.15",
-    "rxjs": "^6.5.2",
-    "rxjs-compat": "^6.5.2",
-    "skrollr-typed": "^0.6.33",
-    "steal": "^2.2.1",
-    "swiper": "^4.5.0",
-    "systemjs": "^5.0.0",
-    "tslib": "^1.9.3",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.9.1"
+    "lodash": "^4.17.15",
+    "ngx-quill": "^7.3.1",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "tslib": "^1.10.0",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.802.0",
-    "@angular/cli": "^8.1.1",
-    "@angular/compiler-cli": "^8.1.2",
-    "@angular/language-service": "^8.1.2",
+    "@angular-devkit/build-angular": "^0.803.4",
+    "@angular/cli": "^8.3.4",
+    "@angular/compiler-cli": "^8.2.7",
+    "@angular/language-service": "^8.2.7",
     "@babel/parser": "^7.4.5",
     "@babel/types": "^7.4.4",
     "@types/jasmine": "^3.3.13",
@@ -170,8 +126,8 @@
     "stylelint-config-standard": "^18.0.0",
     "terser": "^4.0.0",
     "ts-node": "^8.3.0",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3",
+    "tslint": "^5.20.0",
+    "typescript": "^3.6.3",
     "vinyl-paths": "^3.0.0"
   }
 }
