@@ -31,10 +31,6 @@
     // Relative Paths
     paths: {
 
-      /* Require.js Plugins */
-      // This is used to load raw TEXT in templates (e.g. via require.js) -- in most cases we use Angular
-      text: boot.bundle + 'stratus/node_modules/requirejs-text/text',
-
       /* Stratus Core Library */
       stratus: boot.bundle + 'stratus/dist/stratus' + boot.suffix,
 
@@ -59,9 +55,6 @@
       'stratus.directives.base': boot.bundle + 'stratus/directives/base' + boot.suffix,
 
       // THIRD PARTY: NODE MODULES
-
-      /* Underscore is used in most components */
-      underscore: boot.bundle + 'stratus/node_modules/underscore/underscore' + boot.dashSuffix,
 
       /* Lodash is used in place of Underscore in most modern components */
       lodash: boot.bundle + 'stratus/node_modules/lodash/lodash' + boot.suffix,
@@ -167,16 +160,16 @@
       // Angular Dependencies
       'core-js/*': boot.bundle + 'stratus/node_modules/core-js/*',
       'core-js/es7/reflect': boot.bundle + 'stratus/node_modules/core-js/proposals/reflect-metadata',
-      'hammerjs': boot.bundle + 'stratus/node_modules/hammerjs/hammer' + boot.suffix,
-      'rxjs': boot.bundle + 'stratus/node_modules/rxjs/bundles/rxjs.umd' + boot.suffix,
+      hammerjs: boot.bundle + 'stratus/node_modules/hammerjs/hammer' + boot.suffix,
+      rxjs: boot.bundle + 'stratus/node_modules/rxjs/bundles/rxjs.umd' + boot.suffix,
       'rxjs/operators': boot.bundle + 'stratus/extras/normalizers/rxjs.operators' + boot.suffix,
       'rxjs-compat': boot.bundle + 'stratus/node_modules/rxjs-compat/index',
       'web-animations-js': boot.bundle + 'stratus/node_modules/web-animations-js/web-animations.min',
       'zone.js/dist/zone': boot.bundle + 'stratus/node_modules/zone.js/dist/zone' + boot.suffix,
 
       // Quill Editor Support
-      'quill': boot.bundle + 'stratus/node_modules/quill/dist/quill',
-      'ngx-quill': boot.bundle + 'stratus/node_modules/ngx-quill/bundles/ngx-quill.umd',
+      quill: boot.bundle + 'stratus/node_modules/quill/dist/quill',
+      'ngx-quill': boot.bundle + 'stratus/node_modules/ngx-quill/bundles/ngx-quill.umd'
 
       // Angular
 
@@ -205,10 +198,6 @@
       /* STRATUS EXTRAS: Masonry */
       // 'masonry-native': boot.bundle + 'stratus/node_modules/masonry-layout/dist/masonry.pkgd' + boot.suffix,
       // masonry: boot.bundle + 'stratus/extras/directives/masonry' + boot.suffix,
-
-      /* STRATUS Libraries: CodeMirror */
-      codemirror: boot.bundle + 'stratus/node_modules/codemirror/lib/codemirror',
-      'codemirror/*': boot.bundle + 'stratus/node_modules/codemirror/*'
 
     }
   })
