@@ -35,7 +35,7 @@
   Stratus.Filters.Age = function () {
     return function (birthday) {
       birthday = new Date(new Date(birthday).getTime() * 1000)
-      let today = new Date()
+      const today = new Date()
       let age = (today - birthday) / 31557600000
       age = Math.floor(age)
       return age
