@@ -1,19 +1,19 @@
-import { Injectable } from '@angular/core';
-import {HttpClient, HttpResponse} from '@angular/common/http';
-import {Observable} from 'rxjs';
+import { Injectable } from '@angular/core'
+import {HttpClient, HttpResponse} from '@angular/common/http'
+import {Observable} from 'rxjs'
 
 // External
-import * as Stratus from 'stratus';
-import * as _ from 'lodash';
+import * as Stratus from 'stratus'
+import * as _ from 'lodash'
 
 // Interfaces
 export interface Query {
-    fetched: number;
-    data: Observable<HttpResponse<any>>;
+    fetched: number
+    data: Observable<HttpResponse<any>>
 }
 
 export interface QueryMap {
-    [key: string]: Query;
+    [key: string]: Query
 }
 
 @Injectable({
@@ -22,31 +22,31 @@ export interface QueryMap {
 export class BackendService {
 
     // Basic Component Settings
-    title = 'tree-dnd';
-    uid: string;
+    title = 'tree-dnd'
+    uid: string
 
     // HTTP Logic
-    cache: QueryMap;
+    cache: QueryMap
 
     constructor(private http: HttpClient) {
         // Initialization
-        this.uid = _.uniqueId('sa_backend_service_');
-        Stratus.Instances[this.uid] = this;
+        this.uid = _.uniqueId('sa_backend_service_')
+        Stratus.Instances[this.uid] = this
 
         // Instantiate Cache
-        this.cache = {};
+        this.cache = {}
     }
 
     get(url: string): Observable<HttpResponse<any>> {
         if (!_.has(this.cache, url)) {
-            const now = new Date();
-            const data = this.http.get(url, { observe: 'response' });
+            const now = new Date()
+            const data = this.http.get(url, { observe: 'response' })
             this.cache[url] = {
                 fetched: now.valueOf(),
                 data
-            };
-            return data;
+            }
+            return data
         }
-        return this.cache[url].data;
+        return this.cache[url].data
     }
 }
