@@ -1,9 +1,9 @@
-import {ChangeDetectionStrategy, Component, Injectable} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Injectable} from '@angular/core'
 
-import * as Stratus from 'stratus';
-import * as _ from 'lodash';
+import * as Stratus from 'stratus'
+import * as _ from 'lodash'
 
-const localDir = '/assets/1/0/bundles/sitetheorystratus/stratus/src/angular';
+const localDir = '/assets/1/0/bundles/sitetheorystratus/stratus/src/angular'
 
 /**
  * @title Basic Load
@@ -19,6 +19,6 @@ const localDir = '/assets/1/0/bundles/sitetheorystratus/stratus/src/angular';
 // @Injectable()
 export class BaseComponent {
     constructor() {
-        Stratus.Instances[_.uniqueId('sa_base_component_')] = this;
+        Stratus.Instances[_.uniqueId('sa_base_component_')] = this
     }
 }
