@@ -1,11 +1,11 @@
-import {A11yModule} from '@angular/cdk/a11y';
-import {DragDropModule} from '@angular/cdk/drag-drop';
-import {PortalModule} from '@angular/cdk/portal';
-import {ScrollingModule} from '@angular/cdk/scrolling';
-import {CdkStepperModule} from '@angular/cdk/stepper';
-import {CdkTableModule} from '@angular/cdk/table';
-import {CdkTreeModule} from '@angular/cdk/tree';
-import {NgModule} from '@angular/core';
+import {A11yModule} from '@angular/cdk/a11y'
+import {DragDropModule} from '@angular/cdk/drag-drop'
+import {PortalModule} from '@angular/cdk/portal'
+import {ScrollingModule} from '@angular/cdk/scrolling'
+import {CdkStepperModule} from '@angular/cdk/stepper'
+import {CdkTableModule} from '@angular/cdk/table'
+import {CdkTreeModule} from '@angular/cdk/tree'
+import {NgModule} from '@angular/core'
 import {
     MatAutocompleteModule,
     MatBadgeModule,
@@ -41,9 +41,9 @@ import {
     MatToolbarModule,
     MatTooltipModule,
     MatTreeModule,
-} from '@angular/material';
+} from '@angular/material'
 
-import {MatDialogModule} from '@angular/material/dialog';
+import {MatDialogModule} from '@angular/material/dialog'
 
 @NgModule({
     exports: [
