@@ -1,5 +1,5 @@
 // import 'core-js/es6/reflect';
-import 'core-js/es7/reflect';
-import 'zone.js/dist/zone';
-import 'hammerjs';
-import 'web-animations-js';
+import 'core-js/es7/reflect'
+import 'zone.js/dist/zone'
+import 'hammerjs'
+import 'web-animations-js'
