@@ -3,16 +3,16 @@
 // import * as bowser from 'bowser';
 
 // Modules
-import {Environment} from '@stratus/core/environment';
-import {DOM} from '@stratus/core/dom';
-import {Conversion} from '@stratus/core/conversion';
+import {Environment} from '@stratus/core/environment'
+import {DOM} from '@stratus/core/dom'
+import {Conversion} from '@stratus/core/conversion'
 
 export class Stratus {
-    Conversion = new Conversion();
-    DOM = new DOM();
-    Environment = new Environment();
+    Conversion = new Conversion()
+    DOM = new DOM()
+    Environment = new Environment()
     constructor() {
         // console.log('Stratus:', this, _, jQuery, bowser);
-        console.log('Stratus:', this);
+        console.log('Stratus:', this)
     }
 }
