@@ -1,15 +1,16 @@
-let allTestFiles = []
-let TEST_REGEXP = /(spec|test)\.js$/i
+const allTestFiles = []
+const TEST_REGEXP = /(spec|test)\.js$/i
 
 // Get a list of all the test files to include
-Object.keys(window.__karma__.files).forEach(function (file) {
-  if (TEST_REGEXP.test(file)) {
-    // Normalize paths to RequireJS module names.
-    // If you require sub-dependencies of test files to be loaded as-is (requiring file extension)
-    // then do not normalize the paths
-    let normalizedTestModule = file.replace(/^\/base\/|\.js$/g, '')
-    allTestFiles.push(normalizedTestModule)
+Object.keys(window.__karma__.files).forEach(file => {
+  if (!TEST_REGEXP.test(file)) {
+    return
   }
+  // Normalize paths to RequireJS module names.
+  // If you require sub-dependencies of test files to be loaded as-is (requiring file extension)
+  // then do not normalize the paths
+  const normalizedTestModule = file.replace(/^\/base\/|\.js$/g, '')
+  allTestFiles.push(normalizedTestModule)
 })
 
 require.config({
